initial refactor, not working
