+++ AUTHORS
@@ -1,0 +1,2 @@
+Torsten Seemann <xxxx>
+Andrew J. Page <ap13@sanger.ac.uk>
