+++ Makefile.am
@@ -1,0 +1,2 @@
+SUBDIRS=src
+ACLOCAL_AMFLAGS = -I m4
