+++ tests/data/good.aln
