+++ tests/data/gzip.aln.gz
