+++ tests/data/singleton.aln
